backend/requirements.txt
fastapi
uvicorn
openai
supabase
httpx
beautifulsoup4
python-dotenv 